Refactor a few slow rendering fields on the organization form.
