+++ facts/admin/organizations.py
@@ -65,9 +65,7 @@ class PersonShareholderInline(admin.TabularInline):
 class OrganizationAdmin(PhraseSearchAdminMixin, MPTTModelAdmin):
     form = OrganizationForm
     filter_horizontal = [
-        'leaders',
         'initiatives',
-        'related_organizations',
         'related_events',
         'documents',
     ]
